+++ cmd/internal/billing/logic/logic.go
@@ -11,14 +11,14 @@ import (
 	"strings"
 	"text/template"
 
-	"github.com/jmacd/caspar.water/cmd/billing/internal/account"
-	"github.com/jmacd/caspar.water/cmd/billing/internal/business"
-	"github.com/jmacd/caspar.water/cmd/billing/internal/constant"
-	"github.com/jmacd/caspar.water/cmd/billing/internal/csv"
-	"github.com/jmacd/caspar.water/cmd/billing/internal/currency"
-	"github.com/jmacd/caspar.water/cmd/billing/internal/expense"
-	"github.com/jmacd/caspar.water/cmd/billing/internal/payment"
-	"github.com/jmacd/caspar.water/cmd/billing/internal/user"
+	"github.com/jmacd/caspar.water/cmd/internal/billing/account"
+	"github.com/jmacd/caspar.water/cmd/internal/billing/business"
+	"github.com/jmacd/caspar.water/cmd/internal/billing/constant"
+	"github.com/jmacd/caspar.water/cmd/internal/billing/csv"
+	"github.com/jmacd/caspar.water/cmd/internal/billing/currency"
+	"github.com/jmacd/caspar.water/cmd/internal/billing/expense"
+	"github.com/jmacd/caspar.water/cmd/internal/billing/payment"
+	"github.com/jmacd/caspar.water/cmd/internal/billing/user"
 	"github.com/jmacd/maroto/pkg/color"
 	"github.com/jmacd/maroto/pkg/consts"
 	"github.com/jmacd/maroto/pkg/pdf"
